api/requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
psycopg2-binary==2.7.4
PyJWT==1.5.3
pytz==2018.3
